Refactor/monitor rpcs (#337)
Rpc stream servuces refactor:

* refactor monitoring RPCs

fix operations monitor rpc - add protocol and better error handling

* remove client and admin-client binaries from repo

* Remove unused fn and fields

* Fix clippy warnings

* Rework error handling + remove unnecessary unwraps + use ProtcolHash insted of String for protocol field

* Cleanup unnecessary clones

* cleanup unnecessary Options, destructure struct insted of self.

* Fields pub -> private, destruct RpcServiceEnvironment to reduce cloning
