+++ rpc/src/lib.rs
@@ -40,7 +40,7 @@ pub(crate) fn make_json_response<T: serde::Serialize>(content: &T) -> ServiceRes
 }
 
 /// Function to generate JSON response from a stream
-pub(crate) fn make_json_stream_response<T: futures::Stream<Item=Result<String, serde_json::Error>> + Send + 'static>(content: T) -> ServiceResult {
+pub(crate) fn make_json_stream_response<T: futures::Stream<Item=Result<String, failure::Error>> + Send + 'static>(content: T) -> ServiceResult {
     Ok(Response::builder()
         .header(hyper::header::CONTENT_TYPE, "application/json")
         .header(hyper::header::ACCESS_CONTROL_ALLOW_ORIGIN, "*")
