+++ rpc/src/server/shell_handler.rs
@@ -24,9 +24,9 @@ use crate::{
     ServiceResult,
     services,
 };
-use crate::helpers::{create_rpc_request, MempoolOperationsQuery, parse_block_hash, parse_chain_id};
+use crate::helpers::{create_rpc_request, parse_block_hash, parse_chain_id};
 use crate::server::{HasSingleValue, HResult, Params, Query, RpcServiceEnvironment};
-use crate::services::base_services;
+use crate::services::{base_services, stream_services};
 
 #[derive(Serialize)]
 pub struct ErrorMessage {
@@ -69,8 +69,18 @@ pub async fn valid_blocks(_: Request<Body>, _: Params, _: Query, _: RpcServiceEn
 
 pub async fn head_chain(_: Request<Body>, params: Params, query: Query, env: RpcServiceEnvironment) -> ServiceResult {
     let chain_id = parse_chain_id(params.get_str("chain_id").unwrap(), &env)?;
-    let protocol = query.get_str("next_protocol");
-    make_json_stream_response(base_services::get_current_head_monitor_header(&chain_id, &env, protocol.map(|s| s.to_string()))?.unwrap())
+    let protocol = if let Some(protocol) = query.get_str("next_protocol") {
+        HashType::ProtocolHash.b58check_to_hash(protocol).ok()
+    } else {
+        None
+    };
+
+    let RpcServiceEnvironment {
+        state,
+        ..
+    } = env;
+
+    make_json_stream_response(stream_services::HeadMonitorStream::new(chain_id, state, protocol))
 }
 
 pub async fn mempool_monitor_operations(_: Request<Body>, params: Params, query: Query, env: RpcServiceEnvironment) -> ServiceResult {
@@ -81,14 +91,21 @@ pub async fn mempool_monitor_operations(_: Request<Body>, params: Params, query:
     let branch_delayed = query.get_str("branch_delayed");
     let refused = query.get_str("refused");
 
-    let mempool_query = MempoolOperationsQuery{
+    let mempool_query = stream_services::MempoolOperationsQuery{
         applied: applied == Some("yes"),
         branch_refused: branch_refused == Some("yes"),
         branch_delayed: branch_delayed == Some("yes"),
         refused: refused == Some("yes"),
     };
 
-    make_json_stream_response(base_services::get_operations_monitor(&chain_id, &env, Some(mempool_query))?.unwrap())
+    let RpcServiceEnvironment {
+        state,
+        log,
+        ..
+    } = env;
+
+    let last_checked_head = state.read().unwrap().current_head().as_ref().unwrap().header().hash.clone();
+    make_json_stream_response(stream_services::OperationMonitorStream::new(chain_id, state, log, last_checked_head, mempool_query))
 }
 
 pub async fn blocks(_: Request<Body>, params: Params, query: Query, env: RpcServiceEnvironment) -> ServiceResult {
@@ -152,7 +169,7 @@ pub async fn chains_block_id_metadata(_: Request<Body>, params: Params, _: Query
     let chain_id = parse_chain_id(params.get_str("chain_id").unwrap(), &env)?;
     let block_hash = parse_block_hash(&chain_id, params.get_str("block_id").unwrap(), &env)?;
 
-    result_option_to_json_response(base_services::get_block_metadata(&chain_id, &block_hash, &env).map(|res| res), env.log())
+    result_option_to_json_response(base_services::get_block_metadata(&chain_id, &block_hash, &env), env.log())
 }
 
 pub async fn context_raw_bytes(_: Request<Body>, params: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
@@ -172,13 +189,9 @@ pub async fn context_raw_bytes(_: Request<Body>, params: Params, _: Query, env: 
 
 pub async fn mempool_pending_operations(_: Request<Body>, params: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
     let chain_id = parse_chain_id(params.get_str("chain_id").unwrap(), &env)?;
-    result_to_json_response(
-        services::mempool_services::get_pending_operations(
-            &chain_id,
-            env.state(),
-        ),
-        env.log(),
-    )
+    let (pending_operations, _) =
+        services::mempool_services::get_pending_operations(&chain_id, env.state())?;
+    result_to_json_response(Ok(pending_operations), env.log())
 }
 
 pub async fn inject_operation(req: Request<Body>, _: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
@@ -316,13 +329,11 @@ pub async fn preapply_block(req: Request<Body>, params: Params, _: Query, env: R
     match services::protocol::preapply_block(chain_id_param, chain_id, block_hash, rpc_request, &env) {
         Ok(resp) => result_to_json_response(Ok(resp), env.log()),
         Err(e) => {
-            if let Some(err) = e.as_fail().downcast_ref::<ProtocolServiceError>() {
-                if let ProtocolServiceError::ProtocolError { reason: ProtocolError::ProtocolRpcError { reason: ProtocolRpcError::FailedToCallProtocolRpc(message) } } = err {
-                    return make_json_response(&ErrorMessage {
-                        error_type: "ocaml".to_string(),
-                        message: message.to_string(),
-                    });
-                }
+            if let Some(ProtocolServiceError::ProtocolError { reason: ProtocolError::ProtocolRpcError { reason: ProtocolRpcError::FailedToCallProtocolRpc(message) } }) = e.as_fail().downcast_ref::<ProtocolServiceError>() {
+                return make_json_response(&ErrorMessage {
+                    error_type: "ocaml".to_string(),
+                    message: message.to_string(),
+                });
             }
             empty()
         }
@@ -331,7 +342,7 @@ pub async fn preapply_block(req: Request<Body>, params: Params, _: Query, env: R
 
 pub async fn node_version(_: Request<Body>, _: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
     result_to_json_response(
-        base_services::get_node_version(env.network_version()),
+        Ok(base_services::get_node_version(env.network_version())),
         env.log(),
     )
 }
@@ -353,7 +364,7 @@ pub async fn config_user_activated_protocol_overrides(_: Request<Body>, _: Param
 // TODO: remove. This is a 'fake it till you make it' handler
 /// Handler mockin the describe routes in ocaml to be compatible with tezoses python test framework
 pub async fn describe(method: Method, req: Request<Body>, _: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
-    let path: Vec<String> = req.uri().path().split("/").skip(2).map(|v| v.to_string()).collect();
+    let path: Vec<String> = req.uri().path().split('/').skip(2).map(|v| v.to_string()).collect();
 
     let service_fields = serde_json::json!({
         "meth": method.as_str(),
@@ -406,7 +417,7 @@ pub async fn describe(method: Method, req: Request<Body>, _: Params, _: Query, e
 
 pub async fn worker_prevalidators(_: Request<Body>, _: Params, _: Query, env: RpcServiceEnvironment) -> ServiceResult {
     result_to_json_response(
-        base_services::get_prevalidators(&env),
+        Ok(base_services::get_prevalidators(&env)),
         env.log(),
     )
 }
