+++ rpc/src/services/mod.rs
@@ -8,3 +8,4 @@ pub mod dev_services;
 pub mod mempool_services;
 pub mod protocol;
 pub mod stats_services;
+pub mod stream_services;
